+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 
+## 0.7.0
+- **Added**: An option to evaluate configuration without creating an instance of `OpenVPNAdapter`.
+- **Updated**: Class `OpenVPNProperties` renamed to `OpenVPNConfigurationEvaluation`.
+- **Updated**: openvpn3 library to 3.5.6 version, mbedTLS library to 2.7.13 version;
+
 ## 0.6.0
 - **Updated**: Slightly changed API of the framework.
 - **Fixed**: Reading packets issue affecting on connection when network interface is changed.
 
 ## 0.5.0
 - **Added**: Swift Package Manager support;
-- **Updated**: openvpn3 library – 3.5.4 version;
+- **Updated**: openvpn3 library to 3.5.4 version;
 - **Fixed**: Network issue when adapter used in macOS projects.